chore: refactor to move section dropdown to separate component
