+++ src/app/components/NavBar.tsx
@@ -1,10 +1,11 @@
 'use client';
 
 import Link from 'next/link';
-import { Fragment, useState } from 'react';
+import { useState } from 'react';
 import { themes, useThemeContext } from '../theme';
-import { Doc, DocsStructure } from '../types/types';
+import { DocsStructure } from '../types/types';
 import styles from './NavBar.module.css';
+import SectionDropdown from './SectionDropdown';
 
 function VersionsDropdown({
   version,
@@ -91,61 +92,6 @@ function VersionsDropdown({
           ))}
       </div>
     </>
-  );
-}
-
-function SectionDropdownGroup({ doc }: { doc: Doc }) {
-  return (
-    <Fragment>
-      {doc.copy && doc.example && (
-        <option
-          className={styles.selectItem}
-          label={doc.title}
-        >
-          {doc.title}
-        </option>
-      )}
-      {doc.children && doc.children.length > 0 && (
-        <optgroup label={doc.title}>
-          {doc.children.map(child => (
-            <option
-              className={styles.selectItem}
-              label={child.title}
-              key={child.title}
-            >
-              {child.title}
-            </option>
-          ))}
-        </optgroup>
-      )}
-    </Fragment>
-  );
-}
-
-function SectionDropdown({
-  version,
-}: {
-  version: DocsStructure;
-}) {
-  const handleSelectChange = (e: any) => {
-    const selectedOption = e.target.value;
-    location.hash = selectedOption
-      .replaceAll(' ', '-')
-      .toLowerCase();
-  };
-
-  return (
-    <select
-      onChange={handleSelectChange}
-      title="navigation dropdown"
-      className={styles.dropdownSelect}
-    >
-      {version.docs.map(doc => {
-        return (
-          <SectionDropdownGroup doc={doc} key={doc.title} />
-        );
-      })}
-    </select>
   );
 }
